rooms middleware to handle study room requests
