+++ views/pages/study-rooms/study-room-calendar.ejs
@@ -20,7 +20,7 @@
             var roomNumber = window.location.pathname.substr(window.location.pathname.length - 1);
 
             $.ajax({
-                url: "http://localhost:3000/tanulo/" + roomNumber + "/event",
+                url: "http://localhost:3000/tanuloszobak/" + roomNumber + "/event",
                 type: "GET",
                 data: {
                     id: roomNumber
