+++ views/pages/study-rooms/study-rooms.ejs
@@ -18,7 +18,7 @@
                 role="alert"
                 style="margin-bottom: 5px;">
                     <strong><%= i %>. Emelet</strong>
-                    <a href="tanulo/<%= i %>" class="btn btn-default pull-right" style="position: relative; top: -7px; margin: 0 2px; ">Naptár</a>
+                    <a href="tanuloszobak/<%= i %>" class="btn btn-default pull-right" style="position: relative; top: -7px; margin: 0 2px; ">Naptár</a>
                     <% if(color === "green") { %>
                         <a href="csoportok/uj/?roomId=<%= i %>" class="btn btn-success pull-right" style="position: relative; top: -7px; margin: 0 2px; width: 100px">Foglalás</a>
                     <% } %>
@@ -26,7 +26,7 @@
                         <a href="csoportok/<%= groups.find(group => group.room === i).id %>" class="btn btn-warning pull-right" style="position: relative; top: -7px; margin: 0 2px; width: 100px">Részletek</a>
                     <% } %>
                     <% if(color === "red") { %>
-                        <a href="tanulo/<%= i %>" class="btn btn-danger disabled pull-right" style="position: relative; top: -7px; margin: 0 2px; width: 100px">Foglalt</a>
+                        <a href="tanuloszobak/<%= i %>" class="btn btn-danger disabled pull-right" style="position: relative; top: -7px; margin: 0 2px; width: 100px">Foglalt</a>
                     <% } %>
                 </div>
                 <% } %>
